requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
Pillow>=9.5.0
pyyaml>=6.0
tensorboard>=2.13.0
tqdm>=4.65.0
open3d>=0.17.0
scikit-learn>=1.2.0
einops>=0.6.1
timm>=0.9.0
modal>=0.55.0
wandb>=0.15.0
h5py>=3.8.0
pandas>=2.0.0
nuscenes-devkit>=1.1.11

